Fix parquet optional in datafusion-common
